--- src/main/java/GoodbyeCommand.java
@@ -1,11 +1,0 @@
-public class GoodbyeCommand extends Command {
-
-    public GoodbyeCommand(){
-        super(true);
-    }
-
-    @Override
-    void execute(TaskList list, Ui ui, Storage storage) {
-        ui.goodbye();
-    }
-}
