+++ src/main/java/talkgpt/command/Command.java
@@ -1,3 +1,9 @@
+package talkgpt.command;
+
+import talkgpt.storage.Storage;
+import talkgpt.tasklist.TaskList;
+import talkgpt.ui.Ui;
+
 abstract public class Command {
     private boolean isExit;
 
@@ -9,5 +15,5 @@ abstract public class Command {
         return this.isExit;
     }
 
-    abstract void execute(TaskList list, Ui ui, Storage storage);
+    abstract public void execute(TaskList list, Ui ui, Storage storage);
 }
