+++ src/main/java/talkgpt/command/ListCommand.java
@@ -1,0 +1,17 @@
+package talkgpt.command;
+
+import talkgpt.storage.Storage;
+import talkgpt.tasklist.TaskList;
+import talkgpt.ui.Ui;
+
+public class ListCommand extends Command {
+
+    public ListCommand(){
+        super(false);
+    }
+
+    @Override
+    public void execute(TaskList list, Ui ui, Storage storage) {
+        ui.listView(list);
+    }
+}
