+++ src/main/java/talkgpt/parser/Parser.java
@@ -1,3 +1,8 @@
+package talkgpt.parser;
+
+import talkgpt.TalkGPTException;
+import talkgpt.command.*;
+import talkgpt.task.*;
 import java.util.Objects;
 import java.util.Set;
