+++ src/main/java/talkgpt/tasklist/TaskList.java
@@ -1,3 +1,7 @@
+package talkgpt.tasklist;
+
+import talkgpt.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList extends ArrayList<Task> {
